Workaround for #33 - go sub-modules
Fixes: #33

Go sub-modules do not work with this template, because the of
limitations on how local imports and the "replace" word works
in Go modules. You can only use "replace" in the root-level module
i.e. "main.go"

https://github.com/golang/go/wiki/Modules#gomod

This patch was tested with the following repo:

git@github.com:alexellis/golang-middleware-relative-import.git

* Documentation in the README has been updated
* Tested e2e and with a local "go test" in GOPATH with modules
turned on and Go 1.13
* Updated copy commands in Dockerfile so that they use a faster
mechanism, vs creating a new container step to chown (must be
tested with buildkit before merging, since buildkit did not like
this approach in the past when running as a restricted user at
runtime)

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>