Allow for shorter graceful shutdown
The healthcheck_interval env-var allows a user to specify a
shorter time before shutting down the HTTP server to new
connections.

This change has been tested upstream in the of-watchdog which
wraps this template.

The default behaviour is to continue to use the write_timeout
for a safe shutdown duration.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>